+++ app/screens/Home/Index.js
@@ -1,15 +1,10 @@
 /* eslint-disable react/jsx-filename-extension */
 import * as React from 'react';
-import {Text, StyleSheet, View, ImageBackground} from 'react-native';
+import {Text, StyleSheet, View} from 'react-native';
 
 import Menu from '../../components/Menu';
 
 const styles = StyleSheet.create({
-  background: {
-    with: '100%',
-    height: '100%',
-    backgroundColor: '#fff',
-  },
   view: {
     flex: 1,
     justifyContent: 'center',
@@ -34,12 +29,10 @@ const HomeScreen = props => {
   const {navigation} = props;
 
   return (
-    <ImageBackground style={styles.background}>
       <View style={styles.view}>
         <Text style={styles.title}>Cushing</Text>
         <Menu navigation={{navigation}} />
       </View>
-    </ImageBackground>
   );
 };