+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v0.11.0
 
+- [build] React Native from 0.70.6 to 0.71.1
 - [build] Update package.json file.
 
 ## v0.10.0